requirements.txt
PyYAML>=6.0
pyparsing>=3.0.0